+++ src/tests/functional/cfp/test_submission_wizard.py
@@ -9,7 +9,6 @@ from pretalx.submission.models import Submission, SubmissionType
 
 
 class TestWizard:
-
     def get_response_and_url(self, client, url, follow=True, method='POST', data=None):
         if method == 'GET':
             response = client.get(url, follow=follow, data=data)
@@ -26,23 +25,38 @@ class TestWizard:
     def perform_init_wizard(self, client, success=True):
         # Start wizard
         djmail.outbox = []
-        response, current_url = self.get_response_and_url(client, '/test/submit/', method='GET')
+        response, current_url = self.get_response_and_url(
+            client, '/test/submit/', method='GET'
+        )
         assert current_url.endswith('/info/') is success
         return response, current_url
 
     def perform_info_wizard(
-        self, client, response, url, next='questions',
-        title='Submission title', content_locale='en', description='Description',
-        abstract='Abstract', notes='Notes', submission_type=None,
+        self,
+        client,
+        response,
+        url,
+        next='questions',
+        title='Submission title',
+        content_locale='en',
+        description='Description',
+        abstract='Abstract',
+        notes='Notes',
+        submission_type=None,
     ):
         submission_data = {
-            'info-title': title, 'info-content_locale': content_locale,
-            'info-description': description, 'info-abstract': abstract,
-            'info-notes': notes, 'info-submission_type': submission_type,
+            'info-title': title,
+            'info-content_locale': content_locale,
+            'info-description': description,
+            'info-abstract': abstract,
+            'info-notes': notes,
+            'info-submission_type': submission_type,
         }
         key, value = self.get_form_name(response)
         submission_data[key] = value
-        response, current_url = self.get_response_and_url(client, url, data=submission_data)
+        response, current_url = self.get_response_and_url(
+            client, url, data=submission_data
+        )
         assert current_url.endswith(f'/{next}/')
         return response, current_url
 
@@ -53,7 +67,16 @@ class TestWizard:
         assert current_url.endswith(f'/{next}/')
         return response, current_url
 
-    def perform_user_wizard(self, client, response, url, password, next='profile', email=None, register=False):
+    def perform_user_wizard(
+        self,
+        client,
+        response,
+        url,
+        password,
+        next='profile',
+        email=None,
+        register=False,
+    ):
         if register:
             data = {
                 'user-register_email': email,
@@ -61,21 +84,23 @@ class TestWizard:
                 'user-register_password_repeat': password,
             }
         else:
-            data = {
-                'user-login_email': email,
-                'user-login_password': password,
-            }
+            data = {'user-login_email': email, 'user-login_password': password}
         key, value = self.get_form_name(response)
         data[key] = value
         response, current_url = self.get_response_and_url(client, url, data=data)
         assert current_url.endswith(f'/{next}/')
         return response, current_url
 
-    def perform_profile_form(self, client, response, url, name='Jane Doe', bio='l337 hax0r', next='/me/submissions'):
-        data = {
-            'profile-name': name,
-            'profile-biography': bio,
-        }
+    def perform_profile_form(
+        self,
+        client,
+        response,
+        url,
+        name='Jane Doe',
+        bio='l337 hax0r',
+        next='/me/submissions',
+    ):
+        data = {'profile-name': name, 'profile-biography': bio}
         key, value = self.get_form_name(response)
         data[key] = value
         response, current_url = self.get_response_and_url(client, url, data=data)
@@ -86,12 +111,23 @@ class TestWizard:
     def test_wizard_new_user(self, event, question, client):
         event.settings.set('mail_on_new_submission', True)
         submission_type = SubmissionType.objects.filter(event=event).first().pk
-        answer_data = {f'questions-question_{question.pk}': '42', }
+        answer_data = {f'questions-question_{question.pk}': '42'}
 
         response, current_url = self.perform_init_wizard(client)
-        response, current_url = self.perform_info_wizard(client, response, current_url, submission_type=submission_type)
-        response, current_url = self.perform_question_wizard(client, response, current_url, answer_data, next='user')
-        response, current_url = self.perform_user_wizard(client, response, current_url, password='testpassw0rd!', email='testuser@example.org', register=True)
+        response, current_url = self.perform_info_wizard(
+            client, response, current_url, submission_type=submission_type
+        )
+        response, current_url = self.perform_question_wizard(
+            client, response, current_url, answer_data, next='user'
+        )
+        response, current_url = self.perform_user_wizard(
+            client,
+            response,
+            current_url,
+            password='testpassw0rd!',
+            email='testuser@example.org',
+            register=True,
+        )
         response, current_url = self.perform_profile_form(client, response, current_url)
 
         doc = bs4.BeautifulSoup(response.rendered_content, "lxml")
@@ -115,7 +151,16 @@ class TestWizard:
         assert len(djmail.outbox) == 2  # user email plus orga email
 
     @pytest.mark.django_db
-    def test_wizard_existing_user(self, event, client, question, user, speaker_question, choice_question, multiple_choice_question):
+    def test_wizard_existing_user(
+        self,
+        event,
+        client,
+        question,
+        user,
+        speaker_question,
+        choice_question,
+        multiple_choice_question,
+    ):
         submission_type = SubmissionType.objects.filter(event=event).first().pk
         answer_data = {
             f'questions-question_{question.pk}': '42',
@@ -125,9 +170,15 @@ class TestWizard:
         }
 
         response, current_url = self.perform_init_wizard(client)
-        response, current_url = self.perform_info_wizard(client, response, current_url, submission_type=submission_type)
-        response, current_url = self.perform_question_wizard(client, response, current_url, answer_data, next='user')
-        response, current_url = self.perform_user_wizard(client, response, current_url, email=user.email, password='testpassw0rd!')
+        response, current_url = self.perform_info_wizard(
+            client, response, current_url, submission_type=submission_type
+        )
+        response, current_url = self.perform_question_wizard(
+            client, response, current_url, answer_data, next='user'
+        )
+        response, current_url = self.perform_user_wizard(
+            client, response, current_url, email=user.email, password='testpassw0rd!'
+        )
         response, current_url = self.perform_profile_form(client, response, current_url)
 
         doc = bs4.BeautifulSoup(response.rendered_content, "lxml")
@@ -153,14 +204,20 @@ class TestWizard:
         assert len(djmail.outbox) == 1
 
     @pytest.mark.django_db
-    def test_wizard_logged_in_user(self, event, client, question, user):
+    def test_wizard_logged_in_user(
+        self, event, client, question, user, review_question
+    ):
         submission_type = SubmissionType.objects.filter(event=event).first().pk
-        answer_data = {f'questions-question_{question.pk}': '42', }
+        answer_data = {f'questions-question_{question.pk}': '42'}
 
         client.force_login(user)
         response, current_url = self.perform_init_wizard(client)
-        response, current_url = self.perform_info_wizard(client, response, current_url, submission_type=submission_type)
-        response, current_url = self.perform_question_wizard(client, response, current_url, answer_data, next='profile')
+        response, current_url = self.perform_info_wizard(
+            client, response, current_url, submission_type=submission_type
+        )
+        response, current_url = self.perform_question_wizard(
+            client, response, current_url, answer_data, next='profile'
+        )
         response, current_url = self.perform_profile_form(client, response, current_url)
 
         doc = bs4.BeautifulSoup(response.rendered_content, "lxml")
@@ -183,7 +240,13 @@ class TestWizard:
 
         client.force_login(user)
         response, current_url = self.perform_init_wizard(client)
-        response, current_url = self.perform_info_wizard(client, response, current_url, submission_type=submission_type, next='profile')
+        response, current_url = self.perform_info_wizard(
+            client,
+            response,
+            current_url,
+            submission_type=submission_type,
+            next='profile',
+        )
         response, current_url = self.perform_profile_form(client, response, current_url)
 
         doc = bs4.BeautifulSoup(response.rendered_content, "lxml")
@@ -199,15 +262,54 @@ class TestWizard:
         assert len(djmail.outbox) == 1
 
     @pytest.mark.django_db
-    def test_wizard_logged_in_user_no_questions_broken_template(self, event, client, user):
+    def test_wizard_logged_in_user_only_review_questions(
+        self, event, client, user, review_question
+    ):
         submission_type = SubmissionType.objects.filter(event=event).first().pk
 
-        event.ack_template.text = str(event.ack_template.text) + '{name} and {nonexistent}'
+        client.force_login(user)
+        response, current_url = self.perform_init_wizard(client)
+        response, current_url = self.perform_info_wizard(
+            client,
+            response,
+            current_url,
+            submission_type=submission_type,
+            next='profile',
+        )
+        response, current_url = self.perform_profile_form(client, response, current_url)
+
+        doc = bs4.BeautifulSoup(response.rendered_content, "lxml")
+        assert doc.select('.alert-success')
+        assert doc.select('.user-row')
+        sub = Submission.objects.last()
+        assert sub.title == 'Submission title'
+        assert not sub.answers.exists()
+        s_user = sub.speakers.first()
+        assert s_user.pk == user.pk
+        assert s_user.name == 'Jane Doe'
+        assert s_user.profiles.get(event=event).biography == 'l337 hax0r'
+        assert len(djmail.outbox) == 1
+
+    @pytest.mark.django_db
+    def test_wizard_logged_in_user_no_questions_broken_template(
+        self, event, client, user
+    ):
+        submission_type = SubmissionType.objects.filter(event=event).first().pk
+
+        event.ack_template.text = (
+            str(event.ack_template.text) + '{name} and {nonexistent}'
+        )
         event.ack_template.save()
 
         client.force_login(user)
         response, current_url = self.perform_init_wizard(client)
-        response, current_url = self.perform_info_wizard(client, response, current_url, submission_type=submission_type, next='profile')
+        response, current_url = self.perform_info_wizard(
+            client,
+            response,
+            current_url,
+            submission_type=submission_type,
+            next='profile',
+        )
         response, current_url = self.perform_profile_form(client, response, current_url)
 
         doc = bs4.BeautifulSoup(response.rendered_content, "lxml")